Rename lib module to 'MultiVolumeImporterLib'
This will avoid conflict when Slicer issue #3549 (Import scripted module
as python module) will be addressed.
